вынес обработчик сообщений в отдельный класс
