+++ cleintPart/src/main/java/services/ClientHandler.java
@@ -9,10 +9,11 @@ import io.netty.channel.SimpleChannelInboundHandler;
 public class ClientHandler extends SimpleChannelInboundHandler<Object> {
     private String token;
     private String clientFolder = "CLIENT_FOLDER";
-
+MessageProcessing messageProcessing = new MessageProcessing();
     public String getToken() {
         return token;
     }
+;
 
     @Override
     public void channelActive(ChannelHandlerContext ctx) throws Exception {
@@ -28,7 +29,8 @@ public class ClientHandler extends SimpleChannelInboundHandler<Object> {
         ObjectCreatorClass objectCreatorClass = (ObjectCreatorClass) msg;
         String type = objectCreatorClass.getTypeOfMessage();
         System.out.println(type + " received");
-        messageProcessor(type, objectCreatorClass, ctx);
+        messageProcessing.messageProcessor(type, objectCreatorClass, ctx);
+      //  messageProcessor(type, objectCreatorClass, ctx);
     }
 
     @Override
@@ -37,59 +39,24 @@ public class ClientHandler extends SimpleChannelInboundHandler<Object> {
     }
 
 
-    void messageProcessor(String type, ObjectCreatorClass o, ChannelHandlerContext ctx) {
-        if (type.equals("auth_res")) {
-            if (o.getMessage().equals("Authorization success")) {
-                this.token = o.getToken();
-                System.out.println("Authorization success, we have token = " + this.token);
-
-                workingProcess(ctx);
-            } else
-                System.out.println("Authorization failure");
-        }
-
-
-        if (type.equals("giveMeFiles")) {
-            System.out.println("We must send some files to server");
-          //  System.out.println(o.getFileList());
-            for (FileProperty o1 :o.getFileList() ) {
-                System.out.println(o1.getPath());
-                   ObjectCreatorClass file = new ObjectCreatorClass("file", "", o1.getPath());
-                   sendObject(file,ctx);
-            }
-        }
-
-
-    }
-
 
     public void sendObject(ObjectCreatorClass o, ChannelHandlerContext ctx) {
         System.out.println("Sending: " + o.getTypeOfMessage());
-
         ctx.channel().writeAndFlush(o);
-
-
     }
 
 
     public void workingProcess(ChannelHandlerContext ctx) {
 
-
-//        /*Передача фала небольшого файл. работает - пока для теста, чтобы не потерять оставлю тут */
-        //    ObjectCreatorClass file = new ObjectCreatorClass("file", clientFolder, "txt.txt");
-        //   sendObject(file,ctx);
-
-
         //обираем список фалов и хэш клиентской папки, если не задана. если задана то можно будет сохранить все настройки сервера в фаил и оттуда брать. пока по дефолту
         /*создаем струтуру каталогов клиента*/
 
         ObjectCreatorClass tree = new ObjectCreatorClass("tree", clientFolder, "");
-
         sendObject(tree, ctx);
 
-
+        //мониторим папки
+        //Thread fileWatcherThread  =;
     }
-
 
 }
