+++ cleintPart/src/main/java/services/FileWatcher.java
@@ -5,12 +5,15 @@ import java.nio.file.*;
 
 import static java.nio.file.StandardWatchEventKinds.*;
 
+import static java.nio.file.StandardWatchEventKinds.*;
+
 class FileWatcher extends Thread {
 
-/*будем запускать между синхонихациями каталогов*/
+    /*будем запускать между синхонихациями каталогов*/
     @Override
     public void start() {
         try {
+
             WatchService watchService = FileSystems.getDefault().newWatchService();
             Path path = Paths.get("CLIENT_FOLDER");
             path.register(watchService, ENTRY_CREATE, ENTRY_MODIFY, ENTRY_DELETE);
